refactor!: remove deprecated GeneratedVaadinRadioGroup
