+++ vaadin-radio-button-flow-parent/vaadin-radio-button-flow/src/main/java/com/vaadin/flow/component/radiobutton/RadioButtonGroup.java
@@ -22,6 +22,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Stream;
 
+import com.vaadin.flow.component.AbstractSinglePropertyField;
 import com.vaadin.flow.component.AttachEvent;
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.ComponentUtil;
@@ -29,15 +30,19 @@ import com.vaadin.flow.component.DetachEvent;
 import com.vaadin.flow.component.HasHelper;
 import com.vaadin.flow.component.HasLabel;
 import com.vaadin.flow.component.HasSize;
+import com.vaadin.flow.component.HasStyle;
 import com.vaadin.flow.component.HasValidation;
 import com.vaadin.flow.component.ItemLabelGenerator;
+import com.vaadin.flow.component.Tag;
 import com.vaadin.flow.component.UI;
+import com.vaadin.flow.component.dependency.JsModule;
 import com.vaadin.flow.component.dependency.NpmPackage;
-import com.vaadin.flow.component.shared.ClientValidationUtil;
-import com.vaadin.flow.component.shared.HasClientValidation;
-import com.vaadin.flow.component.shared.HasTooltip;
 import com.vaadin.flow.component.radiobutton.dataview.RadioButtonGroupDataView;
 import com.vaadin.flow.component.radiobutton.dataview.RadioButtonGroupListDataView;
+import com.vaadin.flow.component.shared.ClientValidationUtil;
+import com.vaadin.flow.component.shared.HasClientValidation;
+import com.vaadin.flow.component.shared.HasThemeVariant;
+import com.vaadin.flow.component.shared.HasTooltip;
 import com.vaadin.flow.component.shared.ValidationUtil;
 import com.vaadin.flow.data.binder.HasValidator;
 import com.vaadin.flow.data.binder.ValidationStatusChangeEvent;
@@ -67,14 +72,18 @@ import com.vaadin.flow.shared.Registration;
  *
  * @author Vaadin Ltd.
  */
+@Tag("vaadin-radio-group")
+@NpmPackage(value = "@vaadin/polymer-legacy-adapter", version = "24.0.0-alpha6")
+@JsModule("@vaadin/polymer-legacy-adapter/style-modules.js")
 @NpmPackage(value = "@vaadin/radio-group", version = "24.0.0-alpha6")
+@JsModule("@vaadin/radio-group/src/vaadin-radio-group.js")
 public class RadioButtonGroup<T>
-        extends GeneratedVaadinRadioGroup<RadioButtonGroup<T>, T>
-        implements SingleSelect<RadioButtonGroup<T>, T>,
-        HasListDataView<T, RadioButtonGroupListDataView<T>>,
-        HasDataView<T, Void, RadioButtonGroupDataView<T>>, HasValidation,
-        HasHelper, HasSize, HasLabel, HasTooltip, HasValidator<T>,
-        HasClientValidation {
+        extends AbstractSinglePropertyField<RadioButtonGroup<T>, T>
+        implements HasClientValidation,
+        HasDataView<T, Void, RadioButtonGroupDataView<T>>, HasHelper, HasLabel,
+        HasListDataView<T, RadioButtonGroupListDataView<T>>, HasSize, HasStyle,
+        HasThemeVariant<RadioGroupVariant>, HasTooltip, HasValidation,
+        HasValidator<T>, SingleSelect<RadioButtonGroup<T>, T> {
 
     private final KeyMapper<T> keyMapper = new KeyMapper<>();
 
@@ -98,6 +107,9 @@ public class RadioButtonGroup<T>
 
     private static <T> T presentationToModel(
             RadioButtonGroup<T> radioButtonGroup, String presentation) {
+        if (radioButtonGroup.keyMapper == null) {
+            return null;
+        }
         if (!radioButtonGroup.keyMapper.containsKey(presentation)) {
             return null;
         }
@@ -116,8 +128,9 @@ public class RadioButtonGroup<T>
      * Default constructor. Creates an empty radio button group.
      */
     public RadioButtonGroup() {
-        super(null, null, String.class, RadioButtonGroup::presentationToModel,
-                RadioButtonGroup::modelToPresentation, true);
+        super("value", null, String.class,
+                RadioButtonGroup::presentationToModel,
+                RadioButtonGroup::modelToPresentation);
 
         addValueChangeListener(e -> validate());
 
@@ -464,13 +477,13 @@ public class RadioButtonGroup<T>
 
     @Override
     public void setReadOnly(boolean readOnly) {
-        super.setReadonly(readOnly);
+        getElement().setProperty("readonly", readOnly);
         refreshButtons();
     }
 
     @Override
     public boolean isReadOnly() {
-        return super.isReadonlyBoolean();
+        return getElement().getProperty("readonly", false);
     }
 
     /**
@@ -482,9 +495,8 @@ public class RadioButtonGroup<T>
      * @param required
      *            the boolean value to set
      */
-    @Override
     public void setRequired(boolean required) {
-        super.setRequired(required);
+        getElement().setProperty("required", required);
     }
 
     /**
@@ -496,12 +508,18 @@ public class RadioButtonGroup<T>
      * @return the {@code required} property from the webcomponent
      */
     public boolean isRequired() {
-        return super.isRequiredBoolean();
+        return getElement().getProperty("required", false);
     }
 
-    @Override
+    /**
+     * Sets the error message to display when the value is invalid.
+     *
+     * @param errorMessage
+     *            the String value to set
+     */
     public void setErrorMessage(String errorMessage) {
-        super.setErrorMessage(errorMessage);
+        getElement().setProperty("errorMessage",
+                errorMessage == null ? "" : errorMessage);
     }
 
     /**
@@ -509,9 +527,8 @@ public class RadioButtonGroup<T>
      *
      * @return the current error message
      */
-    @Override
     public String getErrorMessage() {
-        return super.getErrorMessageString();
+        return getElement().getProperty("errorMessage");
     }
 
     /**
@@ -520,9 +537,8 @@ public class RadioButtonGroup<T>
      * @param label
      *            value for the {@code label} property in the webcomponent
      */
-    @Override
     public void setLabel(String label) {
-        super.setLabel(label);
+        getElement().setProperty("label", label == null ? "" : label);
     }
 
     /**
@@ -530,19 +546,27 @@ public class RadioButtonGroup<T>
      *
      * @return the {@code label} property from the webcomponent
      */
-    @Override
     public String getLabel() {
-        return super.getLabelString();
+        return getElement().getProperty("label");
     }
 
-    @Override
+    /**
+     * Whether the component has an invalid value or not.
+     *
+     * @return {@code true} for invalid, {@code false} for valid
+     */
     public boolean isInvalid() {
-        return isInvalidBoolean();
+        return getElement().getProperty("invalid", false);
     }
 
-    @Override
+    /**
+     * Sets whether the component has an invalid value or not.
+     *
+     * @param invalid
+     *            {@code true} for invalid, {@code false} for valid
+     */
     public void setInvalid(boolean invalid) {
-        super.setInvalid(invalid);
+        getElement().setProperty("invalid", invalid);
     }
 
     @SuppressWarnings("unchecked")
@@ -675,6 +699,25 @@ public class RadioButtonGroup<T>
         return getItemId(value1).equals(getItemId(value2));
     }
 
+    /**
+     * If true, the user cannot interact with this element.
+     *
+     * @param disabled
+     *            the boolean value to set
+     */
+    protected void setDisabled(boolean disabled) {
+        getElement().setProperty("disabled", disabled);
+    }
+
+    /**
+     * If true, the user cannot interact with this element.
+     *
+     * @return the {@code disabled} property from the webcomponent
+     */
+    protected boolean isDisabledBoolean() {
+        return getElement().getProperty("disabled", false);
+    }
+
     private void updateEnabled(RadioButton<T> button) {
         boolean disabled = isDisabledBoolean()
                 || !getItemEnabledProvider().test(button.getItem());
@@ -698,7 +741,6 @@ public class RadioButtonGroup<T>
         keyMapper.setIdentifierGetter(identifierProvider);
     }
 
-    @Override
     protected void validate() {
         boolean isRequired = isRequiredIndicatorVisible();
         boolean isInvalid = ValidationUtil
