+++ vaadin-radio-button-flow-parent/vaadin-radio-button-flow/src/main/java/com/vaadin/flow/component/radiobutton/RadioGroupVariant.java
@@ -15,12 +15,17 @@
  */
 package com.vaadin.flow.component.radiobutton;
 
+import com.vaadin.flow.component.shared.ThemeVariant;
+
 /**
  * Set of theme variants applicable for {@code vaadin-radio-group} component.
  */
-public enum RadioGroupVariant {
-    LUMO_VERTICAL("vertical"), LUMO_HELPER_ABOVE_FIELD(
-            "helper-above-field"), MATERIAL_VERTICAL("vertical");
+public enum RadioGroupVariant implements ThemeVariant {
+    //@formatter:off
+    LUMO_VERTICAL("vertical"),
+    LUMO_HELPER_ABOVE_FIELD("helper-above-field"),
+    MATERIAL_VERTICAL("vertical");
+    //@formatter:on
 
     private final String variant;
